Stupid ruby_dep dep built into TravisCI
Travis now uses ruby_dep which is an annoying package that will make
all your builds fail because you're testing for outdated Ruby
versions.  I'm against forcing such troubles on people and this writing
is my opinion on the matter.  Code tests should be for compatibility
only.  Security of a language is a system owner's concern and impeding
test suites is a bad move.